WIP: refactoring post processors.
Initial implementation of post processing routines was ad-hoc, at the very
least. Frankly, it was attaching a gizmo to a gun via duct tape. Side effects
could become problem in the future, so whole compilation process had to be
reorganized a bit.

But as a result, we will be able to produce fine-grained postprocs that can do
something actually useful.